alot of stuff

+++ src/controllers/notesController.js
@@ -1,58 +1,73 @@
-import { unlink, readFile } from "fs/promises";
+import { unlink, writeFile, rename } from "fs/promises";
+import { existsSync } from "fs";
 import path from "path";
+import { fileURLToPath } from 'url';
+import { randomBytes } from "crypto";
 
 import Notes from "../models/Notes.js";
 import { userPopulateExcludes } from '../config/mongoConfig.js';
+import { generateMdContent } from "../utils/genAiRes.js";
+import catchError from "../utils/catchError.js";
 
-export const getPublicNotes = async (req, res) => {
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+
+export const getNotes = async (req, res) => {
+  let notes;
   try {
-    const notes = await Notes.find({ 'options.isPublic': true }).populate("creator", userPopulateExcludes);
+    if(req?.query?.creator)
+      notes = await Notes.find({ creator: req.query.creator }).populate("creator", userPopulateExcludes);
+    else
+      notes = await Notes.find({ 'options.isPublic': true }).populate("creator", userPopulateExcludes);
     return res.status(200).json(notes);
   } catch (err) {
-    return res.status(500).json({ error: err.message });
+    return res.status(500).json(catchError(err));
   }
 }
 
 export const createNote = async (req, res) => {
-  const { creator, title, content } = req.body;
+  const { creator, title } = req.body;
+  const randomHex = randomBytes(16).toString('hex');
+  const fileName = randomHex + '.md';
+  const filePath = path.join(__dirname, '..', 'uploads', 'notes', fileName);
   try {
+    await writeFile(filePath, '');
     const createdNote = await Notes.create({ 
       creator: creator,
       title: title,
-      content: content
+      file_content: fileName
     });
     return res.status(201).json({
       message: "Note has been created!",
       createdNote: createdNote
     });
   } catch (err) {
-    return res.status(500).json({ error: err.message });
+    return res.status(500).json(catchError(err));
   }
 }
 
 export const getOneNote = async (req, res) => {
   const noteId = req.params.id;
   try {
-    const note = await Notes.findById(noteId).populate("creator", userPopulateExcludes);
-    return res.status(200).json(note);
+    const foundNote = await Notes.findById(noteId).populate("creator", userPopulateExcludes);
+    if(!foundNote) return res.status(404).json({ error: 'Note not found! '});
+    return res.status(200).json(foundNote);
   } catch (err) {
-    if(err.name === "CastError") return res.status(500).json({ error: 'Invalid note id!' });
-    return res.status(500).json({ error: err.message });
+    return res.status(500).json(catchError(err));
   }
 }
 
 export const updateNote = async (req, res) => {
-  if(!req.body || Object.keys(req.body).length === 0) return res.status(400).json({ error: 'No content to update!' });
+  if(!req.body || Object.keys(req.body).length === 0) 
+    return res.status(400).json({ error: 'No content to update!' });
   const noteId = req.params.id;
   const noteData = req.body;
   try {
     const foundNote = await Notes.findById(noteId);
     if(!foundNote) return res.status(404).json({ error: 'Note not found!' });
-    if(foundNote.options?.imported && noteData?.content)
-      return res.status(400).json({ error: 'Cannot edit contents of imported note!' });
 
     if(noteData?.title) foundNote.title = noteData.title;
-    if(noteData?.content) foundNote.content = noteData.content;
+    if(noteData?.accesses) foundNote.accesses = noteData.accesses;
     if(noteData?.options){
       const options = noteData.options;
       if(options?.isPublic) foundNote.options.isPublic = options.isPublic;
@@ -64,7 +79,7 @@ export const updateNote = async (req, res) => {
       updatedNote: updatedNote
     });
   } catch (err) {
-    return res.status(500).json({ error: err.message });
+    return res.status(500).json(catchError(err));
   }
 }
 
@@ -73,51 +88,58 @@ export const deleteNote = async (req, res) => {
   try {
     const deletedNote = await Notes.findByIdAndDelete(noteId);
     if(!deletedNote) return res.status(404).json({ error: "Note not found!" });
-    if(deletedNote.options.imported){
-      const fileName = deletedNote.content;
-      const filePath = path.join(process.cwd(), 'src', 'uploads', 'imported_notes', fileName)
-      await unlink(filePath);
-    }
+    const fileName = deletedNote.file_content;
+    const filePath = path.join(__dirname, '..', 'uploads', 'notes', fileName);
+    await unlink(filePath);
     return res.sendStatus(204);
   } catch (err) {
-    return res.status(500).json({ error: err.message });
+    return res.status(500).json(catchError(err));
   }
 }
 
 export const importNote = async (req, res) => {
   const { creator, title } = req.body;
-  const uploaded = req.file;
+  const filePath = req.file.path;
+  const mimeType = req.file.mimetype;
+
+  const randomHex = randomBytes(16).toString('hex');
+  const fileName = randomHex + '.md';
+  const newFilePath = path.join(__dirname, '..', 'uploads', 'notes', fileName);
   try {
-    // need to make content into md
+    const importedName = path.basename(filePath);
+
+    if(path.extname(importedName) !== '.md'){
+      const mdResult = await generateMdContent(filePath, mimeType);
+      try {
+        const jsonResult = JSON.parse(mdResult);
+        await unlink(filePath);
+        return res.status(400).json(jsonResult);
+      } catch (e) {
+        await writeFile(newFilePath, mdResult);
+        await unlink(filePath);
+      }
+    } else {
+      await rename(filePath, newFilePath);
+    }
+
     const createdNote = await Notes.create({
       creator: creator,
       title: title,
-      content: uploaded.filename
+      file_content: fileName,
     });
+
     return res.status(201).json({
       message: "Note has been imported!",
       createdNote: createdNote
     });
   } catch (err) {
     try {
-      await unlink(uploaded.path);
+      if(existsSync(filePath)) await unlink(filePath);
+      if(existsSync(newFilePath)) await unlink(newFilePath);
     } catch (unlinkErr) {
-      return res.status(500).json({ 
-        error: err.message + ' ' + unlinkErr.message
-      });
-    } 
-    return res.status(500).json({ error: err.message });
-  }
-}
-
-export const getImportedNote = async (req, res) => {
-  const noteFile = req.params.noteFile;
-  try {
-    const filePath = path.join(process.cwd(), 'src', 'uploads', 'imported_notes', noteFile);
-    await readFile(filePath);
-    return res.status(200).sendFile(filePath);
-  } catch (err) {
-    if(err.name === "CastError") return res.status(500).json({ error: 'Invalid note id!' });
-    return res.status(500).json({ error: err.message });
+      const errs = { message: err.message + ' CANNOT UNLINK: ' + unlinkErr.message };
+      return res.status(500).json(catchError(errs));
+    }
+    return res.status(500).json(catchError(err));
   }
 }
